refactor: Rename to stars
