+++ lib/subatomic/icons/starsSolid.js
